requirements.txt
openai
langchain-community
streamlit
